refactor: cria arquivo de utils para desenho e modifica nome de arquivo 'parede' para 'labirinto'
