+++ src/labirinto/labirinto.py
